+++ rust/zkgroup/src/api/call_links/params.rs
@@ -33,10 +33,13 @@ pub struct CallLinkPublicParams {
 }
 
 impl CallLinkSecretParams {
+    const ROOT_KEY_MAX_BYTES_FOR_SHO: usize = 16;
+
     pub fn derive_from_root_key(root_key: &[u8]) -> Self {
+        let byte_count = Self::ROOT_KEY_MAX_BYTES_FOR_SHO.min(root_key.len());
         let mut sho = Sho::new(
             b"Signal_ZKGroup_20230419_CallLinkSecretParams_DeriveFromRootKey",
-            root_key,
+            &root_key[..byte_count],
         );
         let uid_enc_key_pair = zkcredential::attributes::KeyPair::derive_from(sho.as_mut());
 
@@ -78,5 +81,39 @@ impl CallLinkSecretParams {
             &ciphertext.ciphertext,
         )?;
         uid.try_into().map_err(|_| ZkGroupVerificationFailure)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::call_links::CallLinkSecretParams;
+
+    #[test]
+    fn test_call_link_secret_params_ignores_extra_bytes() {
+        let bytes_0 = b"0123456789012345";
+        let bytes_1 = b"012345678901234512345";
+
+        assert_eq!(
+            bytes_0.len(),
+            CallLinkSecretParams::ROOT_KEY_MAX_BYTES_FOR_SHO
+        );
+
+        assert!(bytes_1.len() > CallLinkSecretParams::ROOT_KEY_MAX_BYTES_FOR_SHO);
+
+        let secret_params_0 = CallLinkSecretParams::derive_from_root_key(bytes_0);
+        let secret_params_1 = CallLinkSecretParams::derive_from_root_key(bytes_1);
+
+        assert_eq!(
+            secret_params_0.uid_enc_key_pair.a1,
+            secret_params_1.uid_enc_key_pair.a1
+        );
+        assert_eq!(
+            secret_params_0.uid_enc_key_pair.a2,
+            secret_params_1.uid_enc_key_pair.a2
+        );
+        assert!(
+            secret_params_0.uid_enc_key_pair.public_key
+                == secret_params_1.uid_enc_key_pair.public_key
+        );
     }
 }