+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.86.13
 
+* Limit the number of call link root key bytes used for CallLinkSecretParams derivation.